0717 타임리프, 사용자 인증

+++ Chap06/src/main/java/edu/pnu/controller/BoardController.java
@@ -6,13 +6,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import edu.pnu.domain.Board;
+import edu.pnu.domain.Member2;
 import edu.pnu.service.BoardService;
 
-
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 	@Autowired
@@ -37,45 +41,90 @@ public class BoardController {
 //		
 //	}
 	
-	@GetMapping("/getBoardList")
-	public String getBoardList(Model model, Board board) {
-		List<Board> boardList = boardService.getBoardList(board);
+	
+	@GetMapping("/helloty")
+	public String helloty(Model model) {
+		model.addAttribute("helloty", "Hello_타임리프");
+		return "hello";
+	}
+	
+	@ModelAttribute("member")
+	public Member2 setMember() {
+		return new Member2();
+	}
+	
+	@RequestMapping(value="/getBoardList", method= {RequestMethod.GET, RequestMethod.POST})
+	public String getBoardList(@ModelAttribute("member")Member2 member, Model model, Board board) {
 		
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
+		List<Board> boardList = boardService.getBoardList(board);
 		model.addAttribute("boardList", boardList);
 		return "getBoardList";
 		
 	}
+//	@RequestMapping(value="/getBoardList", method= {RequestMethod.GET, RequestMethod.POST})
+//	public String getBoardList(Model model, Board board) {
+//		
+//	
+//		List<Board> boardList = boardService.getBoardList(board);
+//		model.addAttribute("boardList", boardList);
+//		return "getBoardList";
+//		
+//	}
+	
 	
 	@GetMapping("/insertBoard")
-	public String insertBoardView() {
-		
+	public String insertBoardView(@ModelAttribute("member")Member2 member) {
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
 		return "insertBoard";
 		
 	}
 	//게시글 등록 처리
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member")Member2 member,Board board) {
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
+		
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 		
 	}
 	@GetMapping("/getBoard")
-	public String getBoard(Board board , Model model) {
+	public String getBoard(@ModelAttribute("member")Member2 member,Board board , Model model) {
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
+		
 		model.addAttribute("board",boardService.getBoard(board) );
 		return "getBoard";
 		
 	}
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member")Member2 member,Board board) {
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.updateBoard(board);
 		return "forward:getBoardList";
 		
 	}
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member")Member2 member,Board board) {
+		if (member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 		
 	}
+	
+		
+		
+	
 }
